extract benches::util module
